refactor: global에 init data 추가
